requirements.txt
python-telegram-bot!=21.3,>=21.0
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.10
sqlalchemy==2.0.31
httpx>=0.27.0
python-dotenv==1.0.1
